+++ process/tests/pipeline.rs
@@ -1,18 +1,16 @@
-use process::{Command, Error};
+#[cfg(feature = "async-std")]
+use async_std::test;
+use process::{Error, Pipeline};
+#[cfg(feature = "tokio")]
+use tokio::test;
 
-#[tokio::test]
+#[test_log::test(test)]
 async fn test_pipeline() {
-    env_logger::builder().is_test(true).init();
-
-    let cmd = Command::from(vec!["echo hello", "cat"]);
+    let cmd = Pipeline::new(vec!["echo hello", "cat"]);
     let out = cmd.run().await.unwrap().to_string_lossy();
     assert_eq!(out, "hello\n");
 
-    let cmd = Command::from("echo hello | cat");
-    let out = cmd.run().await.unwrap().to_string_lossy();
-    assert_eq!(out, "hello\n");
-
-    let cmd = Command::from(vec!["echo hello", "bad", "cat"]);
+    let cmd = Pipeline::new(vec!["echo hello", "bad", "cat"]);
     match cmd.run().await.unwrap_err() {
         Error::GetExitStatusCodeNonZeroError(cmd, status, err) => {
             assert_eq!(cmd, "bad");